Изменен тип данных свойства ИД единицы работ с byte на short как в БД.

+++ RM/RM.Api/DTOs/Requests/WorkTypeCreationRequest.cs
@@ -13,5 +13,5 @@ public class WorkTypeCreationRequest
     /// <summary>
     /// Получает или задает ИД единицы работ для вида работ.
     /// </summary>
-    public byte? WorkUnitId { get; set; }
+    public short? WorkUnitId { get; set; }
 }
