+++ RM/RM.BLL.Abstractions/Models/WorkTypeCreationModel.cs
@@ -13,6 +13,6 @@ public class WorkTypeCreationModel
     /// <summary>
     /// Получает или задает ИД единицы работ для вида работ.
     /// </summary>
-    public byte? WorkUnitId { get; set; }
+    public short? WorkUnitId { get; set; }
 
 }
