+++ RM/RM.DAL.Abstractions/Entities/WorkTypeShortEntity.cs
@@ -20,6 +20,6 @@ public class WorkTypeShortEntity
     /// <summary>
     /// Получает или задает ИД единицы работ для вида работ.
     /// </summary>
-    public byte? WorkUnitId { get; set; }
+    public short? WorkUnitId { get; set; }
 
 }
